requirements.txt
aiogram==3.22.0
aiosqlite==0.21.0
alembic==1.16.4
cachetools==6.2.0
python-dotenv==1.1.1
SQLAlchemy==2.0.43
